Updated and optimized algorithms from TextureAtlas
The insertion algorithm has the logarithmic complexity to insert the object in the dictionary and constant complexity to insert the respective images in the texture atlas.
The removal algorithm has logarithmic complexity to search for an item and constant complexity to remove images from the texture atlas.
Several tests were carried out on 32x32 images and all were successful.