Update of class ImgPack [significant]
The insertion algorithm regressed to linear time (the algorithm was
logarithm).
The removal algorithm regressed to polynomial time O(n^2) (the algorithm
was logarithm (O(log n) + O(1)).
The regression was performed because it requieres the implementation of
a particular data strucutre (A tree with together with a doubly linked
list[dld]). This strcuture was built using std::map plus private dld but
as it is not possible to store the pointer inside the map class, was
necessary implement the current data strcuture wich is temporary.