Added functionality: searchTByName using specifier
It's posible to search for a Thing by specifying a specific type to which it belongs. For example, there can be two things with equal names and distintic types: thing x type ground and thing x type creature. To search for the x thing of type ground, just use the following command: "x,ground" and type enter.