Update MapUI class \*/SIGNIFICANT\*/
The technique used to draw the map structure were changed. Before we were having bad performance with map strucutre with size proximity of 500,500 units Thing objects and was very dificult to control the camera position.
Now we are using the system similar to opengl, apliying model view transformation matrix.
The work is in progress.